Spelling fixes, remove compiler warnings.

+++ src/shaders/ColorDistanceShader2.cpp
@@ -197,7 +197,7 @@ void ColorDistanceShader2::loadState(const JSONObject & dataStore)
 
     for(size_t i = 0; i < 32; ++i)
     {
-        sprintf(buf, "attributeColor_%i", i);
+        sprintf(buf, "attributeColor_%i", (int)i);
 
         setVec4(stringToWstring(buf), colors[i]);
     }
@@ -209,7 +209,7 @@ void ColorDistanceShader2::saveState(JSONObject & dataStore)
 
     for(size_t i = 0; i < 32; ++i)
     {
-        sprintf(buf, "attributeColor_%i", i);
+        sprintf(buf, "attributeColor_%i", (int)i);
 
         dataStore[stringToWstring(buf)] = new JSONValue(colors[i]);
     }
